Point Stripe Events to active pledge form instance
There is a bug with the NewPledgeComponent where the Stripe Element's
events were only being connected to the first instance created, and not
being linked to any subsequent instances after the first is destroyed.
This made it so you sometimes couldn't submit the donation form after
opening and closing it twice. Fix the bug by keeping track of an active
instance in a static variable alongside the static Stripe Element
variable.

PER-8774: Stripe form sometimes cannot submit