Reset PledgeService after making a donation
The PledgeService doesn't properly purge all of its data after making a
donation, which means if a user donates twice in a row without a
refresh, the second donation will be charged, but there will be an error
when it actually comes to claiming storage. Fix this by reseting all of
the PledgeService's data after attempting to make a donation.

PER-8774: 2nd donation in a row will take your money but not give storage