Add reCAPTCHA Site Key as optional secret
The reCAPTCHA component requires an API key to use. Rather than forcing
the application to break when an API key isn't specified (which we
really don't need in testing environments), let's define the reCAPTCHA
key as optional, with the idea that if no reCAPTCHA key is specified,
the application just won't show the reCAPTCHA component.

PER-9034: Add reCAPTCHA to sign in page