Show reCAPTCHA before verifying phone numbers
Show the re-captcha component before sending any verification text
messages to the user. For testing purposes, there is a flag on the
VerifyComponent called `showCaptchaForEmail` which allows for enabling
the reCAPTCHA functionality on e-mail as well as phone numbers, which is
just an easier thing to test with. This also allows us to show a CAPTCHA
for e-mail in the future if we feel we need to do so.

The reCAPTCHA will not be displayed if a reCAPTCHA Site Key is not
passed into the .env file.

While running in unit tests, the re-captcha component should not
actually be activated, instead being treated as a dummy custom element.

PER-9034: Add reCAPTCHA to sign in page