Disable folder thumbnails in archive manager
Due to migration changes and session locking on our API requests, folder
thumbnail generation is too slow and we have decided to disable it in
the archive manager right now. The default `folder` icon is used
instead, without sending any API requests. While thumbnails are disabled
in the archive manager, they are still enabled in the public gallery.

Ideally, folder thumbnails would also explicitly be enabled on the share
preview page as well, however, folder thumbnails cannot be fetched from
the front end on this page because of permissions issues (the folder
thumbnail code has to be able to see the contents of the folder, which
can only be done when logged in or in a public folder).

PER-9067: Disable folder thumbnails in archive manager