Restrict tag service cache to tags from the current archive
The TagsService is intended to keep track of all tags that exist in the
current archive of a logged-in user. It updates its cache of tags when a
new item is viewed in case that item has a recently added tag. However,
this update also happens when viewing items in a public archive, and
tags from such items are added to TagsService cache of tags. This will
cause those tags to appear as options when a user subsequently returns
to their own archive and tries to add a tag to an item, even though
those tags don't actually exist in that archive. This commit fixes that
bug by updating TagsService to check the archiveId of tags before adding
them to its cache, and not add them if they don't match the current
archive.