Add Metadata Manager to Archive Settings
Previous commits for the Metadata Manager were written only in the
context of the unit testing environment. This commit adds the Metadata
Manager to the Archive Settings dialog and adjusts a lot of
functionality that wasn't broken in tests but appeared when testing on a
real browser with real data.

PER-9023: Custom Metadata Manager