Fix Scrolling on File Viewer
Using `justify-content: center` does not allow a user to scroll if the
contents of the flexbox are larger than the flexbox itself. This
prevented a scrollbar from showing up on the file viewer, even when
adding `overflow-y: auto` to the container.

Fix this by adding a wrapper element around the file metadata, and use
auto margin to center content instead of using justify-content.

PER-9128: Clean up public archive metadata bar: make scroll work