Fix unclickable breadcrumbs on public archive
The Profile Photo element was set up with some margins that made the
breadcrumb links on public archives unclickable unless you hovered near
the bottom of the link. Fix this by manually offsetting the elements
which be moved around by the negative margins so that they remain on top
and clickable.

PER-8827: "Public" button breadcrumb does not appear clickable