Install Storybook
Storybook allows us to test components isolated from the rest of the
app. We should probably have a utility like this in our project and
start to add our components to it. This commit installs Storybook.

Most of this commit was generated with Storybook's installer. All of the
files in `src/stories` and `.storybook/` were CLI-generated. One of the
only manual changes to these files was adding the `disableTelemetry`
property to Storybook's main configuration.

PER-9180: Add storybook to web-app