Add Default Access Role to Share Links
The back end and mobile app both support adding a default access role to
share links. Add this feature to the web-app.

In addition, create a Story to test this in Storybook, extract out
shared mocks that both unit tests and stories use, and refactor some of
the methods used in the Sharing Dialog Component. Also, Prettier was run
on all code as well.

PER-9142: Add default access role to share links (front end)