PER-9157: Redirect reset password links to FusionAuth
This is in progress, since it still needs to send to either the dev or
prod fusionauth instance. It may also be Too Hacky, waiting on review
for that determination. In any case with a change to the Forgot
Password template in FusionAuth (on the local tenant only), this does
indeed send users to the web app and then on to FusionAuth, so we can
presumably follow that redirect from the mobile apps too.