Add Directives Display Screen Stories
This adds stories for the Directives Display Screen, as well as all of
the corresponding code that goes along with it. There are a few parts of
the design that will be implemented in the future, such as showing a
proper edit button if a directive already exists for the user. Also the
embedded links don't link anywhere right now.

This commit also introduces the Directive model, and the DirectiveRepo
in the API Service. Currently the DirectiveRepo is unimplemented and
will be hooked up to actual API endpoints in a future ticket.

PER-9164: Directives Display Screen