Add Legacy Contact API Stub
The Legacy Contact API is still under construction, but the directive
display UI needs to check for the presence of a legacy contact on the
current account. Stub out a method to fetch the Legacy Contact for an
account and create the type definition for the response.

The `LegacyContact` type is modeled after the Legacy Contact table
schema from the back-end.

PER-9165: Display "No Plan" warning on directive display