Show "No Plan" Warning on Directives Display
In the design for the Directives Display screen, there is a warning that
shows up and prohibits the user from continuing if the current account
does not have a Legacy Contact set up. This commit includes the code to
show that warning.

After discussing with the team, we've decided that this UI
functionality may end up disabled when the feature goes live. As a
result, all of this functionality can be switched on or off with an
input in the `DirectiveDisplayComponent`.

PER-9165: Display "No Plan" warning on directive display