PER-9157: Pass FA host by environment
Thanks to @meisekimiu for guidance on this! I created the
`FUSIONAUTH_PROD_HOST` as a repository secret. It isn't really secret
since it'll appear as the link in every "forgot password" email, so it
could go straight into the `build-prod` workflow as
https://auth.permanent.org. If I have issues with the value being
populated correctly then I may use that instead.