Expect API to return Steward Email in Directive
Previously, the Directive type had a `stewardAccountId` property on it,
with the intention that the UI would be able to look up account
information to fill in the e-mail address of an Archive Steward (and
vice versa when it comes to creating/editing directives). We realize now
that with our current API design and security standards that this is not
possible outside of the mocked environment made for storybook in
previous commits.

This commit updates the Directive model to return a `stewardEmail`
property instead of an account ID. Because of this change, a lot of the
code initially written for the `DirectiveDisplayComponent` can be
simplified or deleted. The component no longer has to call any Account
API endpoints, and as a result, the code looking up account information
and the mocked `AccountRepo` can all be deleted. The component can also
look up the property on the directive object it stores instead of having
to define a new variable.

PER-9164: Directive Display Screen