Create DirectiveEditComponent
This component makes up the screen to edit/create an archive steward for
the current archive. Currently this component is not implemented into
the web-app or a complete flow yet. It can only be tested in isolation
in Storybook or Karma. An upcoming ticket that adds API functionality
will connect the edit and display screens into a proper flow in the
web-app.

PER-9152: Directive edit screen