Upgrade to Angular 15
This commit includes the package.json and lockfile updates made to
upgrade Angular to version 15.

These changes were made from the following command:
```
npx ng update @angular-eslint/builder@15 \
    @angular-eslint/eslint-plugin@15 \
    @angular-eslint/eslint-plugin-template@15 \
    @angular-eslint/schematics@15 \
    @angular-eslint/template-parser@15 \
    @angular/core@15 @angular/cli@15 @angular/cdk@15 \
    @angular/google-maps@15 @angular/compiler-cli@15 \
    @angular/language-service@15 @angular/localize@15 \
    @angular/animations@15 @angular/common@15 \
    @angular/forms@15 @angular/platform-browser@15 \
    @angular/router@15 @ng-bootstrap/ng-bootstrap@14 \
    shallow-render@15 @angular-devkit/build-angular@15 \
    @fortawesome/angular-fontawesome@^0.12 angular-shepherd@15 \
    ngx-cookie-service@15 @angular/fire@^7.5 ng-recaptcha@^11
```

This upgrade bumped the version of TypeScript which generated errors in
code that was using the `vis-data` types. This was fixed with the
following command:
```
npm install --save vis-data@^7.1.6 vis-timeline@7.7.2
```

PER-9245: Upgrade to Angular 15