Explicitly Define PromptService.prompt return type
This was being parsed as a Promise<unknown> under a newer version of
TypeScript, which was breaking compilation as one cannot access
properties or methods of an `unknown` type object with narrowing its
type first.

This commit also includes changes to other PromptService methods to fix
the same issue.

PER-9245: Upgrade to Angular 15