Automated Sass Migration: Strict Unary Operators
Newer versions of Sass are preparing to make the `+` and `-` operators
less ambiguous. While the current version of Sass installed in the
Angular 15 update only show warnings and not errors, it's best to get
rid of the new warning messages now.

This fix was done via an automatic migration. More information can be
found at https://sass-lang.com/documentation/breaking-changes/strict-unary

PER-9245: Upgrade to Angular 15