Clean the variables when a folder is selected
Reset the variables to empty strings when a folder is selected. This was done because the previous variable was saved when a folder was selected and when a file was selected the previous variable was used to display the file type. This was causing the file type to be displayed incorrectly.