Create New HttpV2Service
The HttpService is used to handle the request/response structure of
Permanent's V1 API. As we have begun to implement V2 versions of
endpoints, we initially extended the HttpService with new methods to
call V2 requests. However, we've now begun needing to call endpoints
with more than just the POST method that the V1 Permanent API uses for
each endpoint.

To fix this problem in the most usable way, create a new service that
more clearly separates the code that calls V1 and V2 endpoints. This new
service now has four methods for HTTP verbs: GET, POST, PUT, and DELETE.
It also uses the `responseClass` parameter in a bit more of a type safe
way.

In addition, there is no equivalent to the `sendRequestPromise` method
of the V1 HttpService, as the caller can just call `.toPromise()` on
their own.

PER-9154: Implement Directives API