Update existing API V2 requests and add tests
The `BaseRepo` class is configured to take the HttpV2Service as an
optional parameter so that the repos that currently don't call V2
endpoints don't need their test code updated. Outside of tests, all
repos are properly passed the HttpV2Service.

However, the AccountRepo and AuthRepo both have code that called the old
`sendV2Request` method. Swap these for calls from the new HttpV2Service
and make sure that all of these methods are properly unit tested so we
know if anything breaks in refactoring this code.

PER-9154: Implement Directives API