Replaced the file type
Display the file type from the FileVOs array. From my understanding there are max 2 items in the array, the original one and the converted one. The original format value comes from the first item of the array. If the second item exists display it in the permanent format, otherwise display the original format.