Refactor HttpV2Service
This commit updates HttpV2Service to return arrays instead of singular
objects, and exports the `getFirst` method to convert an Array
Observable to a Singular Observable in the event the caller only wants
the first object returned in the array.

This also makes the HttpV2Service store the AuthToken in LocalStorage
and also allows the service to prevent sending CSRF and AuthToken
responses if needed.

PER-9154: Implement Directives API