PER-9221 Made it possible to choose file type when downloading an item
PER-9221

This pr servers as the implementation of the feature that allows the users to select the file type when downloading the highlighted item. The dropdown appears in the bar above the table, but also in the info side of the selected item.

The dropdown does not appear if there are multiple files selected or if the selected file is a folder.

It also appears in the file viewer coming from the archive or from the public archive.

There seem to be some issues with the downloadURL for some files (ex: docx, because for each converted value the download url is the same, so this might be a backend issue). and also for images coming from the public archive, the same behaviour happening there aswell