Remove shallow-render from TagsService tests
Somehow, tests from the TagsService started to fail. The reason is that
they were using shallow-render which was improperly mocking the
ApiService and not injecting in the HttpClientTestingModule.

This fixes that error. There have been a lot of shallow-render related
issues with unit testing lately and it begs the question: should we
continue to use it? I still think it's valuable for making Angular unit
tests easier to read and write, but it's concerning how often they can
randomly break.