Install country-list package
Instead of relying on a giant JSON file in our codebase, let's use an
NPM package that provides the dataset instead. This commit installs the
country-list package as a dependency of web-app.

Also install @types/country-list as a dev dependency, since country-list
is a JavaScript library with no type definitions.

PER-9256: Inaccurate names in country list