Use country-list in PrConstantsService
Instead of including a JSON file in our own project, use the
country-list package to get a list of country names and their codes
instead.

The country-list package supplies the overwrite function to allow users
to inject in custom data for specific countries. It is used here to
correct some information that is provided in the default country list.

PER-9256: Inaccurate names in country list