Update Directive components for Directive Modal
When the initial Directive components were being made, the `Directive`
type was just an interface. With the Directive API code finally wired
up, the `Directive` interface was renamed to `DirectiveData` with all
existing usage renamed as well. This commit updates the new components
to use the new `Directive` class.

In addition, make some changes to the Directive Display and Edit
components so they can be wired up properly in a modal, including
outputting events, making sure data sent to the API matches what the API
is expecting, and allowing the directive display component to show
cached data before it is refreshed.

PER-9250: Create directive modal