Create DirectiveDialogComponent
This component is the main container that controls the flow between the
DirectiveDisplay and DirectiveEdit components.

Because all functionality lies in the subcomponents, this component does
not have any sophisticated tests. I've kept the "test for existence"
test intact, but no actual functionality lives in this component other
than switching states.

PER-9250: Create directive modal