Add hidden Legacy Planning tab to Archive Settings
This feature is not ready to go completely live yet, but this adds
the ability to load in the Legacy Planning tab on the Archive Settings
page with a URL hack. The dialog subscribes to URL changes so the user
doesn't have to perform a hard refresh after performing the URL hack.

PER-9250: Create directive modal