Created a new component for the search results
I have created a new component for the search results that has it's own route mainly because of the router outler.

I am also passing the query param inside the route now and the fetch action happens in the ngOnInit hook on the new component. This also helps with navigating to specific records or folders.