Directive API now returns Steward object
Update the Directive data types to reflect the new change in the
Directive API on Stela. Instead of returning a `stewardEmail` field,
return a `steward` object that holds both name and email fields. Update
the data types to reflect that as well as the code that references these
data types.

The API currently has a bit more of an asymmetrical shape now (requests
take a `stewardEmail` field while responses return the `steward` object)
so the method to update a Directive object had to have its typing
adjusted. Instead of taking a `Partial<DirectiveData>` type, it now uses
a new `DirectiveUpdateRequest` type instead, which more accurately
represents the data anyway (since not all Directive fields are accepted
by Stela in an update request).

PER-9250: Create directive modal