PER-9267: use show more for long descriptions
For this functionality, I thought that 100 characters for the short text might be enough, but that could change upon request.

When the user clicks the new span with show more, the full text gets displayed, and when the user clicks on see less, the shortened version is being shown.

The short text is a new variable saved inside the component