Create LegacyContactEditComponent
Create the Legacy Contact Edit screen. The template is mainly based on
the DirectiveEditComponent, which is why some adjustments have been made
to the that component's SCSS file.

This commit stubs out the rest of the Legacy Contact API methods in the
ApiService. The methods will be fully implemented in a future ticket.
(PER-9162).

PER-9147: Create Legacy Contact Edit Screen