Log out user on invalid auth token
Currently on our dev instance of FusionAuth, auth tokens only last for
an hour with no way of being refreshed. This leads to tokens for calling
Stela endpoints expiring pretty quickly and before the PHP session
expires, leading to errors in Stela API calls. Fix this by logging out
the user when the Auth Token expires.

This is accomplished by having an event fire off when the HttpV2Service
receives an HTTP 401 status code. The AccountService is subscribed to
this event and will log out the user and refresh the page to boot them
back to the log in screen.

PER-9283: Send users back to login if they don't have a valid JWT