Implement Legacy Contact API
Previously, the Legacy Contact API was stubbed out since it didn't need
to be called for Storybook testing. Implement methods in the
DirectiveRepo of the ApiService that lets us hit the Legacy Contact
endpoints of stela.

Before the API was implemented in Stela, it was assumed that we might
have to pass in the AccountID when calling these endpoints. Ultimately
the real endpoints are able to get the currently authenticated account
so this is unnecessary. As a result, we don't need to pass in an
AccountVO on any Legacy Contact API methods and so previous calls to the
stubbed out API methods that assumed this have been updated to remove
AccountVO parameters.

PER-9162: Implement Legacy Contact API