Add optional STELA_DOMAIN secret
On local environments, the domain to access Stela endpoints is the same
as the standard API. In dev/staging/prod, Stela will live on a different
domain and this domain needs to be configured for these different
environments. Allow the Stela domain to be configured with a secret, and
then make the HttpV2Service use this domain for Stela requests.

PER-9287: Configure web-app with Stela API Domain