PER-9235: archive modal sometimes not opening
When accesing the public archive, then navigating to the private page, and then clicking the archive settings modal, there was an error being thrown.

The public module was missing the resolver for the profile items. Instead the profile items were being initialized using the resolver for the public items.

I have added the missing resolver and renamed the profileItems variable to publicProfileItems.