Include LegacyContactId with update call
Previously, the legacyContactId property was not included with the
updateLegacyContact call. This ID is required (otherwise how is it
supposed to know which legacy contact to update?) but was not included
since it wasn't part of the mocked version made for unit tests. Include
it here now that we're integrating this into the UI.

PER-9251: Build Legacy Contact Modal