Rename settings-dialog to account-settings-dialog
Now that there are both an Account Settings dialog and an Archive
Settings Dialog, the name "settings-dialog" can be a bit confusing.
Refactor the component to be named "account-settings-dialog" to reflect
what it's called in the app and what the Permanent team actual calls the
dialog in tickets.

PER-9251: Build Legacy Contact Modal