Fix link to Legacy Contact Modal
The Archive Steward Display Screen (DirectiveDisplayComponent) features
a link to a user's Legacy Contact. Fix this link so it actually links to
the Legacy Contact modal.

This link causes a "hard" navigate in the app, meaning the page is
actually reloaded. The reason for this is that the web-app has
difficulties navigating between different modals; it seems like the
Dialog system wasn't built to support this. Angular's `routerLink`
directive also does not seem to support escaping from a named router
outlet (i.e. the `(private//dialog:settings)` that can be seen in the
URL when the dialog is open). Bypassing this restriction with a manual
`Router.navigate` call also does not work as for some reason it ends up
creating an inescapable never ending loop between the two modals.

PER-9251: Build Legacy Contact Modal