Save auth token when verifying MFA
Previously, the Auth Token was only recorded on the `login` endpoint,
which fails when MFA is required to log in. Only once a user
successfully verifies through the `verify` endpoint do they get an Auth
Token. Because the verification step also gives a `trustToken`, we have
to manually find the specific instance of SimpleVO or AuthSimpleVO with
the key "authToken" which is probably a better way of looking for the
token than relying on just finding the first SimpleVO in the data set.

Fortunately it looks like the token is grabbed when creating an account
for the first time.

PER-9284: Store JWT in all authentication flows