PER-9237 gift-storage-interface
This pr's purpose is the implementation of the gift storage interface.

I have added a new tab to the dropdown for the gift component. I have created a new component which in which the new form lies where the user must input the email, amount and the message. Upon confirmation a new modal pops up for confirmation.