Show different welcome message for default signups
Show a welcome message reminding users who signed up with no special
invite code to check their email to redeem their storage.

Note that this solution does have an edge case bug: If a user signs up
with a specific promo code but does not complete onboarding and does so
later on another device/browser, they will be shown this new message
instead of the default message, which may be inaccurate. We've accepted
this as a good enough solution for the time being since there's no way
to get a user's invite code via the API right now.

PER-9280: Add message to redeem storage for earlyb1rd signups