PER-9242 - Add payer archive settings UI
PER-9242 - Add payer archive settings UI

I have added a new section to the archive settings tab where a logged user can set themselves as the payer of the archive.

I have created a new component ArchiveStorageComponent which lies in the archive settings. When the user toggles to toggle, a popup appears with a message,
asking the user if they want to set themselves as the payer of the archive. If the user clicks yes, the payer is set to the logged user.

If the user is already the payer, the user can remove themselves as the payer of the archive. Another message appears asking the user if they want to remove
themselves as the payer of the archive. If the user clicks yes, the payer is set to null.