Fix search by tag
Recently, it looks like there was an attempt to switch calls to the
/search/folderAndRecord endpoint over to use V2 requests, which was then
reversed. However, the name of the tags passed in the request body was
changed to "tags" from "TagVOs" and not changed back, which breaks
searching by tag. This commit returns the name of that field to
"TagVOs".