Add eslint ruleset with failing rules disabled
Install our preferred configuration/plugins to match our other
TypeScript repos, and then disable every single rule.

To mark the rules we want to eventually enable, the `.eslintrc.json`
file was converted to the `.eslintrc.js` format instead which allows for
comments. That way we can separate the rules we are temporarily
disabling from the rules we want to eventually enable.

The rules are disabled in alphabetical order with a few exceptions.
Presumably we will begin enabling rules in alphabetical order. To keep
things simple to start with the import ruleset was brought to the top
since these rules are very likely not going to break anything (or if
they do break something, it will cause a build and test failure rather
than a subtle change in behavior).

Also, since we have not run Prettier on each file yet, it has been put
to the bottom of the list.