Add no-console rule to eslint config
This is a handy rule to catch lingering `console.log` statements from
making their way through a pull request. It's also an easy rule to
enable so we can test lint failures in our CI. Enable this and fix the
only `console.log` statement left in the codebase.

PER-8554: Add eslint to CI