Fix default APP_CONFIG export
The one default export in our codebase was for the APP_CONFIG value used
in a few places. Make this a named export and adjust the imports.
Thankfully each import was already refering to the value as APP_CONFIG
so the imports just needed curly brackets around the import name.

Lint rule: import/no-default-export