Enable import/no-extraneous-dependencies linting rule
This linting rule forbids importing modules that are not declared as
specific dependencies of the project. With our current configuration,
the default behavior was also to forbid the imports of devDependencies,
but due to our project configuration (with things like Storybook and
various configuration scripts we have to use with Angular) it doesn't
make sense to forbid importing devDependencies, at least in files that
are only run in the development environment.

Enable the rule with the `devDependencies` property set to true. This
should prevent future imports of extraneous dependencies but silence our
currently existing errors which only happen in development-related code.

Lint rule: import/no-extraneous-dependencies