Refactor MixPanel code and disable during tests
MixPanel was trying to run in tests whenever the MIXPANEL_TOKEN
environment variable was being set. Add a global variable that can be
configured to force MixPanel to be disabled even when the token is
specified. To make it so code does not have to check this global
variable every time, wrap all the MixPanel code in a service that
verifies that MixPanel is enabled before sending requests.

PER-9325: Integrate MixPanel into the web-app