Verify user is logged in before using cached login
The web-app caches a user's account information during a logged in
session, but sometimes their back end session can expire without this
front end cached information expiring as well. The web-app will continue
to assume the user is still logged in and try to fetch the user's data
even though they are not authorized anymore, causing endpoints to error
out, leaving the user in a never ending invalid state until they clear
their localStorage.

Fix this by checking if the user is logged in before proceeding to use
the cached account info, and clear that cache if the user is not logged
in.

PER-9233: Login spins forever