Create new onboarding flow
PER-9228

Implementation of the new onboarding flow screens.

This was supposed to be 3 tickets, but it made more sense to me to make it only 1, otherwise I think it would have gotten a bit messy, having to change things across multiple prs

Besides the new screens, I have also created a new component for the archive type select, which is used in the new onboarding flow, and also in the create new archive screen.