Enable @angular-eslint/component-selector rule
This linting rule enforces a specific prefix for all defined components,
to ensure that they are valid custom HTML elements. Enable the rule and
set the prefix to "pr", which is the standard we use in our codebase.

This rule is disabled in the Storybook directories since they provide
example components with different prefixes.

Lint rule: @angular-eslint/component-selector