Rename app-root to pr-app-root
The @angular-eslint/component-selector linting rule enforces a
consistent prefix for Angular component selectors within a project. The
rest of the project uses the "pr" prefix for components, with the
exception of the "app-root" element. Rename it to "pr-app-root" to make
it consistent with the rest of the project and pass lint checks.

This change also means that the usage of the tag in the various
`index.html` files for each environment needed to be changed. Some
additional code (a CSS rule and a line of TypeScript referring to the
component) needed to be changed as well.

Lint rule: @angular-eslint/component-selector