Configure @angular-eslint/directive-selector rule
Enable the @angular-eslint/directive-selector rule. This rule forces
directives to follow a standard for naming similar to the
component-selector rule. All of our custom directives already conform to
a standard with no outliers, so the only thing to do is to configure
this rule to match the standard of our codebase.

Lint rule: @angular-eslint/directive-selector