Fix @angular-eslint/no-conflicting-lifecycle rule
The only violation of the no-conflicting-lifecycle rule in our codebase
was in the ThumbnailComponent, which implemented both OnChanges and
DoCheck methods. Refactor the component to run both checks in the same
lifecycle method.

In Angular, OnChanges runs when inputs change, and DoCheck runs whenever
change detection *may* run. Since DoCheck is more general, OnChanges was
removed and its logic was incorporated into DoCheck instead. To get
behavior right (and tests passing), I needed to recreate the conditions
for OnChanges within DoCheck, which meant manually varifying when inputs
change inside of the component.

Lint rule: @angular-eslint/no-conflicting-lifecycle