Fix scroll bug after clicking share link dialog
The current dialog system locks the browser from scrolling while a
dialog is active. This caused a bug in the new dialog that pops up on
the Share Link page. Normally dialogs are closed out in the app and can
reset this scrolling lock, but this specific dialog has routerLinks that
navigate away from the page without closing the dialog, thus leaving the
scrolling lock active. This prevents users from completing onboarding if
the page contents need to be scrolled, since they can't click buttons to
progress. Fix this by closing the dialog when either of the links is
clicked.

PER-9350: Cannot see buttons to progress on onboarding screen from a
share