Add tests for SearchService.parseSearchTerm
SearchService previously had no tests. (Or to be more specific, it only
had commented out test boilerplate!) Let's fix that and add some tests.
First off, add tests that cover the SearchService.parseSearchTerm
function. These are grouped into their own describe block since the
function is a bit complex and the tests have their own helper assertion.

Some of the functionality of this function is a bit unexpected
(sometimes it returns `undefined` when really it should probably be
returning an empty string), but I've written the test cases to match
behavior for now. If we want to change this down the line, we'll have
test cases to guard against breaking other parts of service!