Add SearchService.getTagResults tests
These tests are pretty much copy-and-pasted from the
getResultsInCurrentFolder tests since these two methods are essentially
copy-and-pasted themselves. Now that we have written these tests perhaps
we can do something about that duplication...

Also, while we're on the subject of tags, add a new ParseSearchTerm test
to show some extra (probably erroneous) behavior of the SearchService.
We probably want to fix this at one point, but for now capture it in the
tests.