Add access modifiers to methods in SearchService
Adding explicit access modifiers makes it more apparent which methods
are part of the public API and which methods are just used internally.
While the SearchService class was organized in a way where all
internally called methods were moved to the bottom, we have a way to
explicitly define this in TypeScript. External users of SearchService
have no need to know that `indexCurrentFolder` and `indexTags` even
exist, so let's make them hidden.