Re-enable SignupEmbedComponent test
While formatting this spec file with Prettier, I noticed a commented out
test. Commented out code should generally be deleted, but it was clear
that this test was commented out because functionality changed and the
test started failing. Uncomment out the test and change it to represent
the new intended functionality of this component. Also fix up some other
tests which may have been passing with false positives.

Lint rule: @angular-eslint/no-host-metadata-property