Fix usage of HttpClient in tests
The HttpClient shouldn't be imported in tests as this can cause some
unexpected behavior. Angular provides the HttpClientTestingModule to
import into unit tests which provide testing versions of HttpClient and
HttpHandler which should be used instead.

(It should be noted that Angular generated these usages of HttpClient to
begin with, without importing the testing library!)