Refactor StorageDialog tests
Tests were failing since the shallow-render library was automatically
stubbing out our mock (which we didn't want to do!). To fix this, add a
`dontMock` call to AccountService and use a provider instead.

Since these tests are being modified, refactor them a bit. Split the
"update account after submiting the form" into two tests for each of the
spies and assertions, and delete the "should format in MB" delete since
all it was testing was that our mocks were being properly injected into
the component; something that is tested in the rest of our tests.

Also replace the usage of the `spyOn` function with built in
functionality of our testing doubles. We already have a mocked/stubbed
out version of both ApiService and AccountService for these tests, so
it's a bit unusual to create a spy of an already existing testing
double. Convert the mock objects to classes that are instantiated in the
beforeEach block so that tests have direct access to the whole objects
and they are reset for each test.

PER-9349: Storage Redeemed Does Not Show In Storage Bar