Update to karma-coverage package
Update to the new karma-coverage package and configure Karma to use it
instead of the old and deprecated karma-coverage-istanbul-reporter
package. Generate a new karma config for our current version of Angular,
and then copy some of our old configuration into it. Update angular.json
to point to the new location of the Karma config file, which is now at
the root of the project (where it probably should be).

PER-9359: Add test coverage tool to web-app