Increase AnnouncementComponent test coverage
As part of a quarterly goal to raise test coverage, add some tests so
that this PR increases test coverage.

These two tests cover some code that was originally untested in the
AnnouncementComponent. The code adjusts various elements on the page
when the announcement appoint appears and disappears, and this was
originally written as a quick fix that probably wasn't deemed testable
at the time.

However, shallow-render allows us to render additional HTML in the page
besides our component, so we can add in some external divs to be
adjusted by the AnnouncementComponent for testing.